fix 点击 Save，弹 loading 弹窗，点击 back 键，不允许关闭弹窗

+++ app/src/main/java/com/san/audioeditor/viewmodel/AudioCutViewModel.kt
@@ -77,6 +77,7 @@ class AudioCutViewModel(var song: Song) : BaseViewModel<AudioCutPageData>() {
         var song: Song? = null,
         var isShowEditLoading: Boolean? = null,
         var progress: Int? = null,
+        var isEnableBack: Boolean? = null,
     )
 
 
@@ -95,46 +96,38 @@ class AudioCutViewModel(var song: Song) : BaseViewModel<AudioCutPageData>() {
     }
 
     var outputPath = ""
-    fun save(
-        context: Context,
-        realCutPieceFragments: List<CutPieceFragment>?,
-        datas: MutableList<AudioFragmentBean>
-    ) {
+    fun save(context: Context,
+             realCutPieceFragments: List<CutPieceFragment>?,
+             datas: MutableList<AudioFragmentBean>) {
+        isCancel = false
         if (realCutPieceFragments.isNullOrEmpty()) {
-            ToastCompat.makeText(context, false,context.getString(R.string.error_save)).show()
+            ToastCompat.makeText(context, false, context.getString(R.string.error_save)).show()
             return
         }
         mHandler.datas = datas
         mHandler.context = WeakReference(context)
         viewModelScope.launch(Dispatchers.IO) {
-            val PATH =
-                FFmpegApplication.instance?.getExternalFilesDir(Environment.DIRECTORY_MUSIC)?.absolutePath
+            val PATH = FFmpegApplication.instance?.getExternalFilesDir(Environment.DIRECTORY_MUSIC)?.absolutePath
 
             var realCutPieceFragments = realCutPieceFragments?.filter { !it.isFake }
             if (realCutPieceFragments.isNullOrEmpty()) {
                 ToastCompat.makeText(context, context.getString(R.string.error_save)).show()
+                refresh(AudioCutViewModel(isEnableBack = true))
                 return@launch
             }
             var commandLine: Array<String>? = null
             if (!FileUtil.checkFileExist(song.path)) {
+                refresh(AudioCutViewModel(isEnableBack = true))
                 return@launch
             }
             if (!FileUtil.isAudio(song.path)) {
+                refresh(AudioCutViewModel(isEnableBack = true))
                 return@launch
             }
-            outputPath = AudioFileUtils.generateNewFilePath(
-                PATH + File.separator + AudioFileUtils.getFileName(song.path)
-            )
+            outputPath = AudioFileUtils.generateNewFilePath(PATH + File.separator + AudioFileUtils.getFileName(song.path))
 
-            commandLine = FFmpegUtil.cutMultipleAudioSegments(
-                song.path,
-                realCutPieceFragments.toSegmentsArray(),
-                outputPath
-            )
-            android.util.Log.i(
-                BaseAudioEditorView.jni_tag,
-                "outputPath=${outputPath}"
-            ) //打印 commandLine
+            commandLine = FFmpegUtil.cutMultipleAudioSegments(song.path, realCutPieceFragments.toSegmentsArray(), outputPath)
+            android.util.Log.i(BaseAudioEditorView.jni_tag, "outputPath=${outputPath}") //打印 commandLine
             var sb = StringBuilder()
             commandLine?.forEachIndexed { index, s ->
                 sb.append("$s ")
@@ -148,13 +141,7 @@ class AudioCutViewModel(var song: Song) : BaseViewModel<AudioCutPageData>() {
     }
 
     fun getSongInfo(context: Context, songPath: String): Song? {
-        val cursor = context.contentResolver.query(
-            MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
-            null,
-            MediaStore.Audio.Media.DATA + "=?",
-            arrayOf(songPath),
-            null
-        )
+        val cursor = context.contentResolver.query(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, null, MediaStore.Audio.Media.DATA + "=?", arrayOf(songPath), null)
         if (cursor?.moveToNext() == true) {
             return cursor.convertSong()
         }
@@ -177,27 +164,19 @@ class AudioCutViewModel(var song: Song) : BaseViewModel<AudioCutPageData>() {
 
                 FFmpegHandler.MSG_FINISH -> {
                     Log.i(BaseAudioEditorView.jni_tag, "finish resultCode=${msg.obj}")
+                    if(isCancel){
+                        refresh(AudioCutViewModel(isShowEditLoading = false))
+                        return
+                    }
                     if (msg.obj == 0) {
                         refresh(AudioCutViewModel(isShowEditLoading = false))
                         deleteTempFiles()
-                        var resultFilePath =
-                            Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MUSIC).absolutePath + File.separator + AudioFileUtils.getFileName(
-                                oriSong.path
-                            )
-                        var resultFileName = AudioFileUtils.getFileName(
-                            AudioFileUtils.generateNewFilePath(resultFilePath)
-                        )
-                        var file = AudioFileUtils.copyAudioToFileStore(
-                            File(outputPath),
-                            AppProvider.context,
-                            resultFileName
-                        )
+                        var resultFilePath = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MUSIC).absolutePath + File.separator + AudioFileUtils.getFileName(oriSong.path)
+                        var resultFileName = AudioFileUtils.getFileName(AudioFileUtils.generateNewFilePath(resultFilePath))
+                        var file = AudioFileUtils.copyAudioToFileStore(File(outputPath), AppProvider.context, resultFileName)
                         if (file != null) {
                             AudioFileUtils.deleteFile(outputPath)
-                            AudioFileUtils.notifyMediaScanner(
-                                AppProvider.context,
-                                file.absolutePath
-                            ) { path: String, uri: Uri ->
+                            AudioFileUtils.notifyMediaScanner(AppProvider.context, file.absolutePath) { path: String, uri: Uri ->
                                 context?.get()?.let {
                                     val song = getSongInfo(it, path)
                                     if (song != null) {
@@ -212,6 +191,7 @@ class AudioCutViewModel(var song: Song) : BaseViewModel<AudioCutPageData>() {
                                 .show()
                         }
                     }
+                    refresh(AudioCutViewModel(isEnableBack = true))
                 }
 
                 FFmpegHandler.MSG_PROGRESS -> {
